v0.2 release
New pipeline release including correction of small bugs and a new way of calculating the RV drift from the ThAr frames. Now we use a dedicated set of lines (3273 in total) specifically found to create a well-behaved CCF under exposure times below 150s. The precision achieved with this is now around 5 m/s on the ThAr frames.